Change demo app flow
